fix(sleeve-f): close evidence-integrity gaps from PR #3 review
Opus 4.8 + gpt-5.6-terra dual review; fixes implemented by
gpt-5.6-luna, verified against full suite (235 passed).

- replay metadata: declare regime_eligibility residual lookahead in
  both variants; causal report no longer claims lookahead:false
- live parity: gated fields must come from the offline adapter;
  score-only adapters now FAIL with unverified_fields instead of
  auto-passing on snapshot-copied values
- contracts: signature_id now hash-covered; tuple coercion at
  construction (shallow-frozen leak); flock around tail-read+append
  in both writers; golden fixture regenerated
- collectors: raw payload hashes full rows not row counts (dedup
  false-positive); readers verify manifest sha256 fail-closed;
  per-partition sequence race locked
- c2_sets: batch path features grouped per session to match the
  iterator that built the committed matrices; parity test added;
  strict session-order + naive-IST guards
- era_costs: aware datetimes converted to IST before era selection
  (era_table_hash unchanged: 5e1954ce…e74f)
- expired_common: 75-lot bucket for pre-2021-07 expiries
- build_c2_matrices: volatile fields moved to run_report.json so
  hashes.json is byte-reproducible; zero-row guard
- sleeve_f_router: consec streak int8 -> int16 (overflow at 127)
- label parity: float columns tolerance-gated; backfill failure
  report no longer double-counts written days

+++ features/sleeve_f_router.py
@@ -94,13 +94,13 @@ def compute_features(df: pd.DataFrame, trade_date: date) -> pd.DataFrame:
     df["gap_pct"] = (first_open - previous) / max(abs(previous), EPS)
 
     df["up_bar"] = (df["f_close"] > df["f_close"].shift(1)).astype(np.int8)
-    consec = np.zeros(n, dtype=np.int8)
+    consec = np.zeros(n, dtype=np.int16)
     for index in range(1, n):
         if df["up_bar"].iat[index] == 1:
             consec[index] = max(consec[index - 1], 0) + 1
         else:
             consec[index] = min(consec[index - 1], 0) - 1
-    df["consec_bars"] = consec
+    df["consec_bars"] = consec.astype(np.int16)
     df["ema9"] = df["f_close"].ewm(span=9, adjust=False).mean()
     df["ema21"] = df["f_close"].ewm(span=21, adjust=False).mean()
     df["ema_slope"] = (df["ema9"] - df["ema21"]) / df["ema21"].replace(0, np.nan)
